2013.1~rc1 (Grizzly)

+++ cinder/volume/drivers/netapp/iscsi.py
@@ -32,7 +32,7 @@ from suds import client
 from suds.sax import text
 
 from cinder import exception
-from cinder import flags
+from cinder.openstack.common import lockutils
 from cinder.openstack.common import log as logging
 from cinder.volume import driver
 from cinder.volume.drivers.netapp.api import NaApiError
@@ -51,7 +51,8 @@ netapp_opts = [
                help='User name for the DFM/Controller server'),
     cfg.StrOpt('netapp_password',
                default=None,
-               help='Password for the DFM/Controller server'),
+               help='Password for the DFM/Controller server',
+               secret=True),
     cfg.StrOpt('netapp_server_hostname',
                default=None,
                help='Hostname for the DFM/Controller server'),
@@ -77,10 +78,11 @@ netapp_opts = [
                help='Cluster vserver to use for provisioning'),
     cfg.FloatOpt('netapp_size_multiplier',
                  default=1.2,
-                 help='Volume size multiplier to ensure while creation'), ]
-
-FLAGS = flags.FLAGS
-FLAGS.register_opts(netapp_opts)
+                 help='Volume size multiplier to ensure while creation'),
+    cfg.StrOpt('netapp_volume_list',
+               default='',
+               help='Comma separated eligible volumes for provisioning on'
+                    ' 7 mode'), ]
 
 
 class DfmDataset(object):
@@ -108,6 +110,7 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
 
     def __init__(self, *args, **kwargs):
         super(NetAppISCSIDriver, self).__init__(*args, **kwargs)
+        self.configuration.append_config_values(netapp_opts)
         self.discovered_luns = []
         self.discovered_datasets = []
         self.lun_table = {}
@@ -162,10 +165,10 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
         required_flags = ['netapp_wsdl_url', 'netapp_login', 'netapp_password',
                           'netapp_server_hostname', 'netapp_server_port']
         for flag in required_flags:
-            if not getattr(FLAGS, flag, None):
+            if not getattr(self.configuration, flag, None):
                 raise exception.InvalidInput(reason=_('%s is not set') % flag)
-        if not (FLAGS.netapp_storage_service or
-                FLAGS.netapp_storage_service_prefix):
+        if not (self.configuration.netapp_storage_service or
+                self.configuration.netapp_storage_service_prefix):
             raise exception.InvalidInput(
                 reason=_('Either '
                          'netapp_storage_service or '
@@ -181,13 +184,15 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
         """
         self._check_flags()
         self._create_client(
-            wsdl_url=FLAGS.netapp_wsdl_url,
-            login=FLAGS.netapp_login, password=FLAGS.netapp_password,
-            hostname=FLAGS.netapp_server_hostname,
-            port=FLAGS.netapp_server_port, cache=True)
-        self._set_storage_service(FLAGS.netapp_storage_service)
-        self._set_storage_service_prefix(FLAGS.netapp_storage_service_prefix)
-        self._set_vfiler(FLAGS.netapp_vfiler)
+            wsdl_url=self.configuration.netapp_wsdl_url,
+            login=self.configuration.netapp_login,
+            password=self.configuration.netapp_password,
+            hostname=self.configuration.netapp_server_hostname,
+            port=self.configuration.netapp_server_port, cache=True)
+        self._set_storage_service(self.configuration.netapp_storage_service)
+        self._set_storage_service_prefix(
+            self.configuration.netapp_storage_service_prefix)
+        self._set_vfiler(self.configuration.netapp_vfiler)
 
     def check_for_setup_error(self):
         """Check that the driver is working and can communicate.
@@ -383,6 +388,7 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
         self.discovered_datasets.append(ds)
         return ds
 
+    @lockutils.synchronized('netapp_dfm', 'cinder-', True)
     def _provision(self, name, description, project, ss_type, size):
         """Provision a LUN through provisioning manager.
 
@@ -443,6 +449,7 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
             return None
         return volume_type['name']
 
+    @lockutils.synchronized('netapp_dfm', 'cinder-', True)
     def _remove_destroy(self, name, project):
         """Remove the LUN from the dataset, also destroying it.
 
@@ -1146,6 +1153,7 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
 
     def __init__(self, *args, **kwargs):
         super(NetAppCmodeISCSIDriver, self).__init__(*args, **kwargs)
+        self.configuration.append_config_values(netapp_opts)
         self.lun_table = {}
 
     def _create_client(self, **kwargs):
@@ -1170,7 +1178,7 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
         required_flags = ['netapp_wsdl_url', 'netapp_login', 'netapp_password',
                           'netapp_server_hostname', 'netapp_server_port']
         for flag in required_flags:
-            if not getattr(FLAGS, flag, None):
+            if not getattr(self.configuration, flag, None):
                 msg = _('%s is not set') % flag
                 raise exception.InvalidInput(data=msg)
 
@@ -1183,10 +1191,11 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
         """
         self._check_flags()
         self._create_client(
-            wsdl_url=FLAGS.netapp_wsdl_url,
-            login=FLAGS.netapp_login, password=FLAGS.netapp_password,
-            hostname=FLAGS.netapp_server_hostname,
-            port=FLAGS.netapp_server_port, cache=True)
+            wsdl_url=self.configuration.netapp_wsdl_url,
+            login=self.configuration.netapp_login,
+            password=self.configuration.netapp_password,
+            hostname=self.configuration.netapp_server_hostname,
+            port=self.configuration.netapp_server_port, cache=True)
 
     def check_for_setup_error(self):
         """Check that the driver is working and can communicate.
@@ -1237,6 +1246,10 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
         """Driver entry point for destroying existing volumes."""
         name = volume['name']
         handle = self._get_lun_handle(name)
+        if not handle:
+            msg = _("No entry in LUN table for volume %(name)s.")
+            LOG.warn(msg % locals())
+            return
         self.client.service.DestroyLun(Handle=handle)
         LOG.debug(_("Destroyed LUN %s") % handle)
         self.lun_table.pop(name)
@@ -1346,7 +1359,12 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
 
     def delete_snapshot(self, snapshot):
         """Driver entry point for deleting a snapshot."""
-        handle = self._get_lun_handle(snapshot['name'])
+        name = snapshot['name']
+        handle = self._get_lun_handle(name)
+        if not handle:
+            msg = _("No entry in LUN table for snapshot %(name)s.")
+            LOG.warn(msg % locals())
+            return
         self.client.service.DestroyLun(Handle=handle)
         LOG.debug(_("Destroyed LUN %s") % handle)
         self.lun_table.pop(snapshot['name'])
@@ -1487,6 +1505,7 @@ class NetAppDirectISCSIDriver(driver.ISCSIDriver):
 
     def __init__(self, *args, **kwargs):
         super(NetAppDirectISCSIDriver, self).__init__(*args, **kwargs)
+        self.configuration.append_config_values(netapp_opts)
         self.lun_table = {}
 
     def _create_client(self, **kwargs):
@@ -1515,7 +1534,7 @@ class NetAppDirectISCSIDriver(driver.ISCSIDriver):
         """Ensure that the flags we care about are set."""
         required_flags = self.required_flags
         for flag in required_flags:
-            if not getattr(FLAGS, flag, None):
+            if not getattr(self.configuration, flag, None):
                 msg = _('%s is not set') % flag
                 raise exception.InvalidInput(data=msg)
 
@@ -1528,10 +1547,11 @@ class NetAppDirectISCSIDriver(driver.ISCSIDriver):
         """
         self._check_flags()
         self._create_client(
-            transport_type=FLAGS.netapp_transport_type,
-            login=FLAGS.netapp_login, password=FLAGS.netapp_password,
-            hostname=FLAGS.netapp_server_hostname,
-            port=FLAGS.netapp_server_port)
+            transport_type=self.configuration.netapp_transport_type,
+            login=self.configuration.netapp_login,
+            password=self.configuration.netapp_password,
+            hostname=self.configuration.netapp_server_hostname,
+            port=self.configuration.netapp_server_port)
         self._do_custom_setup()
 
     def check_for_setup_error(self):
@@ -1564,6 +1584,10 @@ class NetAppDirectISCSIDriver(driver.ISCSIDriver):
         """Driver entry point for destroying existing volumes."""
         name = volume['name']
         metadata = self._get_lun_attr(name, 'metadata')
+        if not metadata:
+            msg = _("No entry in LUN table for volume/snapshot %(name)s.")
+            LOG.warn(msg % locals())
+            return
         lun_destroy = NaElement.create_node_with_children(
             'lun-destroy',
             **{'path': metadata['Path'],
@@ -1707,7 +1731,8 @@ class NetAppDirectISCSIDriver(driver.ISCSIDriver):
 
     def _create_lun_on_eligible_vol(self, name, size, metadata):
         """Creates an actual lun on filer."""
-        req_size = float(size) * float(FLAGS.netapp_size_multiplier)
+        req_size = float(size) *\
+            float(self.configuration.netapp_size_multiplier)
         volume = self._get_avl_volume_by_size(req_size)
         if not volume:
             msg = _('Failed to get vol with required size for volume: %s')
@@ -1952,7 +1977,7 @@ class NetAppDirectCmodeISCSIDriver(NetAppDirectISCSIDriver):
 
     def _do_custom_setup(self):
         """Does custom setup for ontap cluster."""
-        self.vserver = FLAGS.netapp_vserver
+        self.vserver = self.configuration.netapp_vserver
         # Default values to run first api
         self.client.set_api_version(1, 15)
         (major, minor) = self._get_ontapi_version()
@@ -2243,7 +2268,11 @@ class NetAppDirect7modeISCSIDriver(NetAppDirectISCSIDriver):
 
     def _do_custom_setup(self):
         """Does custom setup depending on the type of filer."""
-        self.vfiler = FLAGS.netapp_vfiler
+        self.vfiler = self.configuration.netapp_vfiler
+        self.volume_list = self.configuration.netapp_volume_list
+        if self.volume_list:
+            self.volume_list = self.volume_list.split(',')
+            self.volume_list = [el.strip() for el in self.volume_list]
         if self.vfiler:
             (major, minor) = self._get_ontapi_version()
             self.client.set_api_version(major, minor)
@@ -2263,8 +2292,12 @@ class NetAppDirect7modeISCSIDriver(NetAppDirectISCSIDriver):
                 avl_vol['block-type'] = vol.get_child_content('block-type')
                 avl_vol['type'] = vol.get_child_content('type')
                 avl_vol['size-available'] = avl_size
-                if self._check_vol_not_root(avl_vol):
-                    return avl_vol
+                if self.volume_list:
+                    if avl_vol['name'] in self.volume_list:
+                        return avl_vol
+                else:
+                    if self._check_vol_not_root(avl_vol):
+                        return avl_vol
         return None
 
     def _check_vol_not_root(self, vol):
@@ -2335,17 +2368,37 @@ class NetAppDirect7modeISCSIDriver(NetAppDirectISCSIDriver):
     def _create_lun_handle(self, metadata):
         """Returns lun handle based on filer type."""
         if self.vfiler:
-            owner = '%s:%s' % (FLAGS.netapp_server_hostname, self.vfiler)
+            owner = '%s:%s' % (self.configuration.netapp_server_hostname,
+                               self.vfiler)
         else:
-            owner = FLAGS.netapp_server_hostname
+            owner = self.configuration.netapp_server_hostname
         return '%s:%s' % (owner, metadata['Path'])
 
     def _get_lun_list(self):
         """Gets the list of luns on filer."""
+        lun_list = []
+        if self.volume_list:
+            for vol in self.volume_list:
+                try:
+                    luns = self._get_vol_luns(vol)
+                    if luns:
+                        lun_list.extend(luns)
+                except NaApiError:
+                    LOG.warn(_("Error finding luns for volume %(vol)s."
+                               " Verify volume exists.") % locals())
+        else:
+            luns = self._get_vol_luns(None)
+            lun_list.extend(luns)
+        self._extract_and_populate_luns(lun_list)
+
+    def _get_vol_luns(self, vol_name):
+        """Gets the luns for a volume."""
         api = NaElement('lun-list-info')
+        if vol_name:
+            api.add_new_child('volume-name', vol_name)
         result = self._invoke_successfully(api, True)
         luns = result.get_child_by_name('luns')
-        self._extract_and_populate_luns(luns.get_children())
+        return luns.get_children()
 
     def _find_mapped_lun_igroup(self, path, initiator, os=None):
         """Find the igroup for mapped lun with initiator."""
